backend/requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
djangorestframework-simplejwt
celery
redis
requests
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
openai==2.8.0
pdfplumber==0.10.3
python-multipart==0.0.6
Pillow==10.1.0
